+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2.17.2] - 2026-03-23
+
+### Fixed
+- Fixed normal irrigation channel recipe
+- Fixed normal/full irrigation channel conversion recipes not properly copying NBT
+
 ## [2.17.1] - 2026-03-16
 
 ### Fixed